Add files via upload
Meha Injector
This project is a tool for injecting DLL files into processes on the Windows platform. It features a simple graphical user interface allowing users to select the process for injection and the DLL file to be injected.

Functionalities:
Users can choose a process from the list of active processes on their computer or manually specify the path to the executable file. They can then select the DLL file to be injected into the chosen process. Additionally, users have the option to select the injection method: Manual Map, Standard, or Automatic.

Upon clicking the "Inject" button, the program initiates the injection process. Upon completion, users will see a message indicating the result - whether the injection was successful or unsuccessful.

Technologies:
The program is written in Python, utilizing the Tkinter library for building the graphical user interface. The ctypes library is employed for interacting with the Windows API. This setup provides simplicity in implementation and usage, as well as flexibility in injecting a wide range of processes.

Note:
This project is intended for educational purposes and can serve as a foundation for further exploration in Python programming and Windows system interaction. Users can extend the functionalities of the program, add new injection methods, or enhance existing ones.

Meha Injector offers a straightforward and intuitive solution to the task of DLL injection, suitable for both programming novices and experienced developers.